mcrouter: double mem limits
Why:
We have observed that mcrouter gets OOMkilled on codfw, this is
a bandaid until we figure out why. This is not observed
on eqiad, even though mw-mcrouter pushes way more traffic.

Note: We lowered the CPU requests a bit, as it may cause issues
in scheduling the pod if a node is under pressure.

Bug: T374025
Change-Id: Iac54af7df5f58181e9a7b271b83ea1894bb7e02a