airflow: enable visualizing logs of DAG runs in the webserver UI
For this to work, the webserver needs to address the scheduler by
its IP and not by pod name (which does not resolve). It also needs
a NetworkPolicy to be able to ingress on the scheduler http port.

Finally, the webserver and scheduler need to share the same secret and
fernet key, which will be injected as env vars from secrets.

Bug: T368737
Change-Id: I888244e46026af51acca85463e44d561cabaf02b