cirrus: Fix cloudelastic saneitizer, and enable private wikis
The recent deploy that changed the `wikis` filter to support
negation turned out to be a partial fix, the filter was parsed
differently in the producer and the consumer but only the
producer side was adjusted to handle negation. This caused the
saneitizer to stop operating due to thinking it had a positive
filter. Updated container fixes issue.

Additionally since the updater now suppors private wikis we can
enable them for consumer-search, leaving it disabled for
consumer-cloudelastic as no private indices exist there.

Change-Id: I8007b0f69f7a74ee5ef28a4f751150321e4ded22