jaeger: set securityContext for the oauth sidecar
It seems that the oauth sidecar doesn't have any security context
to be set, and this leads to a violation when deploying on aux-k8s.
I hoped that using podSecurityContext would have avoided this,
but its spec doesn't allow certain fields like allowPrivilegeEscalation
(that we set), only usable at the container-level.

Once this patch is deployed and working, I'll file a patch to upstream
too to get it included for future releases.

Bug: T369491
Change-Id: Ic1d3b322f8adbb601ba5111ae606d74e0a02be7e