cloudnative-pg-cluster: enable wal upload / backups to s3 by default
The development of this path took a lot of trial and error, documented
in the Phab task itself.

At the core of it, we enable backups, use the AWS/S3 credentials
stored in the private release value file, and the operator takes
care of the rest.

In practice, we had to tweak 2 major things:
- pass the `TMPDIR:/run` env var to the PG pods, to allow barman
  to write its temporary files in a writable directory, as the whole
  filesystem is made readonly by the operator, except for /run
- disable WAL and backup encryption, as bucket encryption is a feature
  of the rados gateway that we have not yet enabled.

Bug: T372281
Change-Id: Idb61884c07f60621c8dc06879cf194c3d89bc7be
Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>