termbox: remove redundant values-test.yaml lines
These are identical in values.yaml, so I don’t think it’s necessary to
override them in values-test.yaml. In general I think it’s desirable if
values-test.yaml is shorter and only contains the differences from
values.yaml that are actually needed (compare values-staging.yaml with
only two lines in it).

Change-Id: Id435920f36a56c6a58f24a51a2ed4ae925952033