airflow: allow the scheduler to be selectively deployed
This will help us migrate the airflow instances to Kubernetes by
migrating the webserver first.

When the scheduler is not deployed to Kubernetes, the following values
need to be set:

```
scheduler:
  remote_host: <ip>
  remote_port: <port>
```

This will automatically create an egress networkpolicy allowing the
webserver to reach out to the remote scheduler.

Note; this also requires that the scheduler has a nftable rule allow
ingress traffic coming from the DSE_KUBEPODS_SUBNET range.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T374936
Change-Id: I4b0dfe6c5ca64a09c8b607cb28e33681b051b351