cirrus: Read from public and private streams
The cirrus updater will be splitting it's updates for private
wikis into their own stream. This mimics how mediawiki is
sending private wiki events in their own stream.

Update the parameters we provide to the updater to match the new
requirements, and update the container to a version that supports
this functionality.

Bug: T374335
Change-Id: I46321a6fa5b46e0c0983ae097a663ca98bd93982