Compute stable secret checksums by default
We change the way secret checksums are computed. We currently checksum
the whole rendered template itself, meaning that the checksum changes
when the chart version is bumped, as the Secret is tagged with
`chart: {{ $.Chart.Name }}-{{ $.Chart.Version }}`

This causes the deployment pods to be rolled out at _each_ new chart
version, even though their own Deployment spec might not have changed at
all, except for the Secret checksum itself, that only changes because the
Secret label changes, due to the chart version bump.

To remediate this, we introduce a `base.helper.resourcesDataChecksum`
helper, that iterates over each rendered resource in a given file
(assumed to be either Secret or ConfigMap resources), and computes a
checksum of the string representation of an array containing the `data`
or `stringData` field of each resources.

This way, the checksum only changes when the *content* of any of the
resurce in the file changes, but stays stable when only the resource
metadata (such as labels) are changed.

Example (I have update the charts/airflow chart to the new templates):
```
\# At this point, we also had upgarded the airflow chart with the new
\# templates
$ git --no-pager diff -u charts/airflow
diff --git a/charts/airflow/Chart.yaml b/charts/airflow/Chart.yaml
index efc2c028..b2d0f4ec 100644

Change-Id: I8a0c2783fa4441fcda20c0715b02383a99770186
--- a/charts/airflow/Chart.yaml
+++ b/charts/airflow/Chart.yaml
@@ -1,4 +1,4 @@
 apiVersion: v2
 description: A Helm chart for Kubernetes
 name: airflow
-version: 0.0.29
+version: 0.0.30

\# This command performs a diff between the chart rendered with the last
\# committed state and with the current changes
$ chart-diff ~/wmf/deployment-charts/charts/airflow -f ~/wmf/deployment-charts/helmfile.d/dse-k8s-services/airflow-test-k8s/values-cloudnative-pg-cluster-dse-k8s-eqiad.yaml -f ~/wmf/deployment-charts/helmfile.d/dse-k8s-services/airflow-test-k8s/values-production.yaml --set environmentName=dse-k8s-eqiad | grep checksum/secrets
$
```

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: I0701c317285f95bf9f56bf9738f09bf9acd0f17b