shellbox: add support for service.deployment: none
When a instance of the shellbox chart sets service.deployment to "none"
no Service objects will be created for that release.

This is meant to be used together with route_via to serve a portion of
traffic via the service associated with another release (e.g., main).

This is a generalization of the 'not (hasSuffix "canary" .Release.Name)'
in numerous other charts, now allowing for any release name (and using a
similar convention as other charts, such as eventstreams).

Bug: T375243
Change-Id: Iacc40c64d1b1376ea6c2cf3b4cdfd2f6c816f0e0