shellbox-syntaxhighlight: add migration release
The "migration" release will be used for a capacity-based migration to
images using PHP 8.1.

A similar pattern will be used for other variants, but syntaxhighlight
is a good starting point, as it's one of the simplest (i.e., low chance
of behavior changes due to dependencies).

Only wired into staging for now.

Bug: T375243
Change-Id: I9b804f60b3e83a516ebbd5d552ce3cfb3e61f36e