canaries: Recreate instead of RollingUpdate
Why:
There is no gain in a rolling update for canaries. They aren't fronted
by any service but rather powering the same service as the main
releases and their size is specifically tuned to be the same as a
standard main release batch of pods. Instead of waiting out to do them
one by one, going for all of them in 1 batch makes sense. There 2 ways
to achieve this. maxSurge and maxUnavailable to 100% and strategy.type
Recreate. The latter is clearer as to why it is chosen, so go for that
one

What:
Switch strategy.type to Recreate. To avoid syntax errors from the
main/global ones, override the rollingUpdate stanzas

Bug: T375477
Change-Id: Ib71d9211ff6acda683aed59f7d201c09035c1c02