airflow: introduce a way to define Airflow variables as values
Variables (https://airflow.apache.org/docs/apache-airflow/stable/howto/variable.html)
are a way for tasks to fetch dynamic values, and, same as connections,
they can be defined in the UI or in a local file.

We will be using these variables in the DAG defined in
https://gitlab.wikimedia.org/repos/data-engineering/airflow-dags/-/merge_requests/844
as a way to define the s3 log retention and the s3 bucket name.

This way, each instance could apply different logging retention
policies, if need be.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T375715
Bug: T375770
Change-Id: Ie26766b5ad0485d0c55d66b90797e6fbd5a04f05