Disable exposure warning for airflow webservers by default
Airflow has a default feature whereby it displays a warning in the UI if
there have been recent requests to /robots.txt in order to help prevent
accidental public exposure.

We have made a conscious decision to make our instances public-facing
and we use OIDC authentication to secure them.

Therefore, we can disable this warning message.

Bug: T375739
Change-Id: I039ab76fa57965a5dfbf2625da2e467c474d8154