cloudnative-pg-cluster: allow the operator to reload a cluster secrets
Cf https://cloudnative-pg.io/documentation/1.24/labels_annotations/

Without this annotation, we needed to restart the operator itself after
we rolled out new s3 keys for a given cluster, for them to be taken into
account. This will allow the operator to take the new keys into account
without having to restart.

The same applies for our CA certificate and the initdb import password.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T375853
Change-Id: Id2941ed64b65f4f589ce1e2f9f57961856669f4e