Release meta_2.0.2 that injects a configuration checksum annotation
The same way we checksumn `Secret` resources defined in the
`secret.yaml` file, we now compute a `checksum/configuration` annotation
which value is a checksum of all the `data` / `stringData` fields of
`ConfigMap` resources defined in `configmap.yaml`.

This way, we automatically provide a way for scaffolded charts to
rolling-restart pods when a secret or configuation changes occur.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: I637655e44ce8ac57d79f70e2ae35406a2cef5638