airflow: Use the analytics user for task executor pods
We now have a mechanism that enables us to run airflow tasks as specific
user accounts, which will help us when we interact with HDFS and
Kerberos.

We wish to test this mechanism by configuring the test-k8s instance to
use the analytics user for its tasks.

Also, we override the docker-registry address to use the discovery
address, since this was omitted.

Bug: T375895
Change-Id: Ic3f4ee9baeb293517505268b60e15ece99a4b61d