wikidata-query-gui: add new service definition
This change adds the new service wikidata-query-gui. The service reuses
the existing miscweb chart. If additional features are needed a dedicated
chart can be used.
Start with one release until it's clear how different query services are
separated and served. We probably need a release for each of the ingress
fqds.

Bug: T350793
Change-Id: I585f3dba7a746bb24c6762d013c66c1b927241d1