mw-(web|api-ext): revert to multi-DC sizing
This reverts mw-web and mw-api-ext to sizing appropriate for multi-DC
serving, based on a combination of sizes prior to the September 2024 DC
switchover and a newer analysis in [0].

Single-DC sizes are left commented with some context on (1) when they
are from (i.e., how potentially stale) and (2) how they were produced.
At a minimum, these should be recalibrated twice per year by subsequent
switchovers.

[0] https://phabricator.wikimedia.org/T376519#10439450

Bug: T376519
Change-Id: I4b22fbf6f7df699dc46f13a6d1adc21342c90d34