mw-script: Add prometheus-statsd-exporter
Why:
Per discussions in #wikimedia-observability and in T375506, some way of
exposing certain metrics from MW that are calculated periodically needs
to exist. Historically the approach was via cron/systemd timers and
Graphite but it's being deprecated, so statsd-export it is.

What:
Follow the exact same pattern as in the other mw-* namespaces and spin
up a statsd exporter to be used by deployments

Bug: T376714
Change-Id: Ife08f79e25d31c597e857bc7e33987f244f3e548