airflow: align default configuration with our pupetized instances
Looking at the Airflow Roles and permissions, we see that only Ops (a
role we don't use) and Admins have the permission to see the
`Admin > Configuration` and have read permission on `/configuration`.

Nevertheless, we only expose non-sensitive configuration (barring any
secret key, db connection string, etc). It is useful to have this
information available, even for admins.

It is also what we currently have in our puppetized instances.

We also
- remove elasticsearch config, that we don't use
- add webserver.instance_name to set a custom title in the browser
  navbar (currently the case for our puppetized instances)
- set `core.default_task_retries` to 5 (currently the case in our
  puppetized instances)
- set `webserver.expose_{hostname,stacktrace}` to true (currently the
  case in our puppetized instances)

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: Iaa43cb0e27304f16ad1c36e140a9e6fa75784d00