mw-debug: Recreate instead of RollingUpdate
Why:
This is an experiment to see if the gains from Recreate are enough to
balance the occasional errors developers using WikimediaDebug extension
will see. The approach has been already tested in 30af6f6a48, however
mw-debug is a different thing. We do expect to see errors here if a
developer uses the WikimediaDebug extension while a deployment is
happening. Whether that's acceptable or not will need to be gauged
versus the whatever speed improvements we end up seeing during
deployments

Instead of waiting out to do them one by one, going for all of them in 1
batch makes sense. There 2 ways to achieve this. maxSurge and
maxUnavailable to 100% and strategy.type Recreate. The latter is clearer
as to why it is chosen, so go for that one

What:
Switch strategy.type to Recreate. To avoid syntax errors from the
main/global ones, override the rollingUpdate stanzas

Bug: T374907
Change-Id: Iceffebc9cd7f409ea28fd709ecc073339bcb3176