echostore: null out certs.kask in env values
Because the base values file is layered before the helmfile-defaults
secrets, doing this in the former is insufficient until I actually
remove the secrets. I do not yet want to do that, in order to facilitate
rollbacks.

Not caught in Ic314e88d8af0e51c64658bf6a5353e3930ca57f1 due to
differences in values evaluation in CI vs. "the real thing".

Bug: T376766
Change-Id: I7616ae466306311c2f67094f978268e303f2ab27