ceph-csi-cephfs: replace the ClusterRole by a list of ns-scoped Roles
This change was requested by elukey, and will ensure we only provide the
ceph-csi-cephfs provisioner Deployment with the minimum set of RBAC
permissions.

This comes at the cost of maintaining a list of namespaces in which our
releases make use of cephfs PersistentVolumes. However, as this is
already standard practice when it comes to the cloudnative-pg-operator,
we can absolutely live with that.

Note that we cannot migrate all permissions to a Role:
- PV are non-namespaced
- VolumeSnapshotClasses are non-namepaced (cf
  https://raw.githubusercontent.com/kubernetes-csi/external-snapshotter/release-5.0/client/config/crd/snapshot.storage.k8s.io_volumesnapshotclasses.yaml)
- StorageClasses are non-namespaced: https://kubernetes.io/docs/concepts/storage/storage-classes/

Bug: T376406
Change-Id: Ie88c4edbb95509067c835b7b2ee172e5d5a93b95