mediawiki: add mercurius features
Add basic config for Mercurius to mediawiki chart. This setup is for
testing a single topic at present as we plan on using scap to manage
our release IDs in future, but for the time being we manually define
the release version on deploy, so we won't be able to avail of the
restart feature.

This config allows us to create a mercurius Job monitoring a single
topic - some codebase changes are needed to allow us to easily tell
mercurius which config file to read on a per-job basis.

Mercurius will first be used with the mw-videoscaler namespace to
replace videoscaling on metal.

Bug: T371701
Depends-on: I37bfabc3bbbff9c6ba390531e491515ee73e6365

Change-Id: I2f35515d4d6d9ea934ca7ee4fc6c181bd30ab67b