airflow-analytic-test: comment out the postgresql deployment
We only want to deploy the webserver for now, so we comment out the
database deployment-related helmfile sections.

We also remove the `config.airflow.postgresPassword` config override,
that would override the real password set in the private value files, as
the release value file has the highest precedence.

We also disable the scheduler from being deployed, as this is only a
webserver migration for now.
We thus specify the host/port of the remote scheduler.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T374948
Change-Id: I5325c038c022f0249f8c68ab371ca29f004cb5ae