airflow: fix missing configmap when the scheduler isn't deployed
We include that configmap 3 times, for each component: webserver,
scheduler and worker (only useful when using KubernetesExecutor) and the
`if` condition was completely wrong.

I decided to remove it altogether instead of having multiple complicated
if blocks, as it only deploys configmaps that end up being unused when
running with an external scheduler or with LocalExecutor.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T374948
Change-Id: I151cac8125da4117d2679e6f70ff6f5dc95695af