shellbox-syntaxhighlight: add "migration" in prod
This adds the migration release in prod environments, but with replicas
scaled to zero (i.e., objects will be present, but no pods exist, and so
no traffic is served on 8.1).

Also shuffles TODOs around a bit in order to point to the correct task.
Updating the httpd-fcgi image will execute independently and be tracked
under T378128.

Once this is verified to work as expected, the same change will be
applied to all other instances (next patch).

Bug: T375243
Change-Id: I01eb83662c16db34cd64fac2e0c25f20565d4ff9