mediawiki: support for service.deployment: none
In order to support introduction of the migration release, which is
intended to route via main in a manner similar to canary, we need to
make the existing "suppress the NodePort service" logic used by the
latter a bit more generic.

Similar to what was done in Iacc40c64d1b1376ea6c2cf3b4cdfd2f6c816f0e0
for shellbox, this uses service.deployment: none to indicate that no
service should be created.

While we technically could remove the special case for "canary" and rely
fully on service.deployment (canary releases already set this), I've not
gone so far as to do so here.

Bug: T377040
Change-Id: Ia9938efb08ef4f69d078d5a4ebfb8db2ddc07d4c