mw-api-int: add migration release
The "migration" release will be used for a capacity-based migration to
mediawiki running on PHP 8.1.

While we expect to use this pattern to incrementally shift internal and
cookie-less external traffic to 8.1 for _all_ large mediawiki instances,
we'll first turn this up for mw-api-int to shake out any issues before
turning up the analogous releases in other namespaces.

Like we did for the directly addressible "next" releases for mw-api-ext
and mw-web in I6f9a31a046f5ffe5892eea1e08ada1daacb31d07:
* Both the servergroup and the mesh tracing service_name (i.e., also the
  envoy local cluster name), are overridden to clearly differentiate the
  migration deployment from main (e.g., when debugging).
* The helmfile-defaults release value files are temporarily overridden
  to point to those for main, until the first scap deployment after is
  it made aware of migration.

Bug: T377040
Change-Id: I3292d098fbad8e7a91c6f4b76333554bc8b808e6