airflow: define a cephfs PVC storing the DAGs
Now that the `ceph-cephfs-ssd` storageclass is available in
dse-k8s-eqiad, we can migrate the emptyDir ephemeral volume for a
persistent cephfs volume.

That volume is mounted in the webserver, scheduler and the worker pods.
(Note: we shpuldn't have to mount it in the webserver pod but we need the
`wmf_airflow_common` directory for unrelated reasons).

We perform the airflow-dags synchronization/pull via a Deployment that runs
alongisde the webserver and scheduler. While this is inherently racy,
the worst case scenario is that the scheduler and webserver crash at
startup because the dags volume is empty.

We no longer peform a `git pull airflow-dags` when starting the worker pods.
This means that while our reliance on gitlab being up is still there, it
is much less stringent. The airflow worker pods should also be a bit
faster to start as well.

This reliance will be fully mitigated when we can rely on the so-called
HDFS-synchronizer.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T368033
Change-Id: I2c58584c7a66d9c0b33131c6477b51051c2cc6c0