ceph-csi-cephs: fix RBAC by granting cluster-wide permisions on PVC and storageclasses
Tailing the logs of the ceph-csi-cephs provisioner logs, I spotted the
following errors:
```
 1 reflector.go:147] k8s.io/client-go/informers/factory.go:159: Failed to watch *v1.PersistentVolumeClaim: failed to list *v1.PersistentVolumeClaim: persistentvolumeclaims is forbidden: User "system:serviceaccount:kube-system:ceph-csi-cephfs-provisioner" cannot list resource "persistentvolumeclaims" in API group "" at the cluster scope
 1 reflector.go:147] k8s.io/client-go/informers/factory.go:159: Failed to watch *v1.StorageClass: failed to list *v1.StorageClass: storageclasses.storage.k8s.io is forbidden: User "system:serviceaccount:kube-system:ceph-csi-cephfs-provisioner" cannot list resource "storageclasses" in API group "storage.k8s.io" at the cluster scope
 ```

This patch hereby grants cluster-wide permissions on PVC and
StorageClass resources to the provisioner.

I also had to add the `kube-system` as part of the tenantNamespaces
list, as the provisioner neeeded to be able to create events and secrets
in the namespace in which it runs.

I also had to rename `.Values.secret.userId` to `.Values.secret.adminId`
(and `userKey` to `adminKey` in the private puppet repo) due to a
discrepancy between the `ceph-csi-rbd` and `ceph-csi-cephfs` charts.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T376406
Change-Id: I87c2698069ccb5d1d407902da0b2d7b7d766abc5