sessionstore: use service mesh in production
As in 706b43626b2dddd7397de552ba10704872d47fd1, use the service mesh
for terminating tls within sessionstore pods. This replicates the
pattern that is seen on echostore and is also already present in
sessionstore staging.

Bump resources to something fairly generous for the mesh pod - this is
probably way too high even with the heavy usage sessionstore sees, but
we can reduce this as needed after our deployment.

We will need to do a `helmfile destroy` for this change as there will
be a resource conflict between ports. Given this and also the general
sensitivity of sessionstore, we will need to do this on a DC-by-DC
basis, depooling as we roll this change out.

Bug: T363996
Change-Id: I66d2d557cf1e60ab99cb546e0a95096df9111810