mediawiki: add remaining migration releases
This extends what I3292d098fbad8e7a91c6f4b76333554bc8b808e6 did for
mw-api-int to the other high-traffic mediawiki instances.

The "migration" release will be used for a capacity-based migration to
mediawiki running on PHP 8.1, applicable to internal and cookie-less
external traffic.

As we did with mw-api-int:
* Both the servergroup and the mesh tracing service_name (i.e., also the
  envoy local cluster name), are overridden to clearly differentiate the
  migration deployment from main (e.g., when debugging).
* The helmfile-defaults release value files are temporarily overridden
  to point to those for main, until the first scap deployment after is
  it made aware of migration.

Bug: T377040
Change-Id: I1c4c60ac30a6532dcaf19d5b896658b5d7e80c4e