Adjust spark-history rediness probe delay
We have been getting regular issues on the yarn UI for spark history.
looking further in to it the issue seems to have stemmed from the rediness_probe on spark-history,
namely "kubelet  Readiness probe failed:"
This patch adjusts the initial delay time to 20s to help relieve some of the alerts and
errors caused by this.

Bug: T378497
Change-Id: I4fc4d8326665d11da416b2291b959da34feeb229