airflow: render the spark/hadoop/hdfs/yarn configuration files
Why
----
Render hadoop, yarn, hive and Spark configuration, so we're able to run
`spark3-submit` processes in the scheduler (when using `LocalExecutor`)
or airflow task pods (when using `KubernetesExecutor`) in cluster mode,
meaning that the Spark driver will be running in YARN, and not in the
pod itself.

What
----
Some of these config files are copied verbatim from an-launcher1002,
and some are rendered, as we might want to set some overrides per env.

I have also tweaked the hadoop log4j.properties file, to make sure all
loggers log to stdout, and not in different log files on disk.

The case of Hadoop, Kerberos & CoreDNS
--------------------------------------
We have had quite a bit of trouble with a combination of Kerberos,
Hadoop, CoreDNS and external_services, so we've had to introduce a
static hostname/IP mapping for the Hadoop masters, rendered in
`/etc/hosts`. When the hadoop library contacts the namenode, it confirms
that its kerberos hostname conforms with the reverse DNS of its IP. This
is usually performed by libkrb5 and can usually be disabled by setting
`rdns = false` in `krb5.conf`, but in our case, this is a hardcoded
behavior in Hadoop itself.

When Hadoop contacts the namenoode, it does so using its .wmnet
hostname. However, the IP of that hostname is included in an "external
services" Service Endpoints, that we use combined with a Calico Network
Policy to allow airflow to egress to all hosts/ports exposed by the hadoop masters.
Because this is defined a Service, we also get a reverse DNS for each
IP, for free:

```
>>> import socket
>>> socket.gethostbyname('hadoop-master-analytics.external-services.svc')
'10.64.36.15'
>>> socket.gethostbyaddr('10.64.36.15')
('10-64-36-15.hadoop-master-analytics.external-services.svc.cluster.local', [], ['10.64.36.15'])
```

By setting `spec.hostAliases` on the pods that will run spark3-submit,
we force the reverse DNS resolution of these IPs to equate to their
wmnet hostanme:
```
airflow@airflow-scheduler-5757cb9dfc-sdz9x:/opt/airflow$ cat /etc/hosts
...
10.64.36.15	an-master1003.eqiad.wmnet
10.64.53.14	an-master1004.eqiad.wmnet
```

```
>>> socket.gethostbyaddr('10.64.36.15')
('an-master1003.eqiad.wmnet', [], ['10.64.36.15'])
```

By doing this, the kerberos hostname validation will run as expected.
See https://phabricator.wikimedia.org/T377602#10295542 for more details.

Additionally, we unify all airflow images to run with the same
`airflow:900:900` runuser, as we've realized that running different
images with different users was:
- causing issues with kerberos token permissions (which was owned by
  airflow:airflow with `o600` permissions
- useless, as the user impersonation didn't rely on the UNIX user but on
  the kerberos principal

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T377928
Bug: T377927
Bug: T379265
Change-Id: Idc1ef5c1923e5be53298f6670fff0b92c0fe8078