mediawiki: ensure default php.version is a string
This was added in I4799e9c2a044488b304d5c63d3ecc15fcdcad3d2, and while
this works as expected (i.e., is in fact a noop), I have concerns about
the way type reflection in go could mess this up upon rendering in the
future.

To be on the safe side, let's make sure this is a string.

Bug: T372604
Bug: T377040
Change-Id: I0b65a55ec02f819cc5fb43d2326058c54cbcf2f0