api|rest-gateway: Support sending request headers to upstream
Why:
For T374683, we need rest-gateway to send a specific header to
MediaWiki, so that MediaWiki replies with RESTBase compatible responses.

What:
Add the ability to specific request headers to be added to the request
towards the upstream cluster, on a per route basis.
Bump the chart to reflect the above
Specify the needed request headers for T374683 to happen

Bug: T374683
Change-Id: I318209234c0a3e848852bec61bd48d20ee1cb0de