changeprop-jobqueue: set max poll interval and revert concurrency
This patch sets max.poll.interval.ms to 1h (default: 5m) for
webVideoTranscode* jobs. This is somewhat arbitrary, but seems like a
reasonable first attempt at preventing frequent partition reassignment
[0].

The temporary concurrency limit reduction in
I70eec970f2a442e30bda0d55e7399b5dcad40384 for webVideoTranscode has
already been reverted, and it's expected that this can / should be
relaxed further if the max.poll.interval.ms tuning works as expected (as
we'll lose some "accidental" concurrency). Also, to some extent, a
tighter limit may make reassignment more frequent.

[0] Largely based on the cpjobqueue_normal_rule_processing_count metric,
which seems to indicate that "at least one completion in a given hour"
is a plausible target. This is of course muddied by the fact that we
overshoot on concurrency due to precisely this problem ...

Bug: T356241
Change-Id: I18907208ac930f40802e60c6fcdaba67f9d2db4c