dse-k8s-services: mw-dump: version bump image
Bump mediawiki-event-enrichment to v1.39.0.
This release fixes a parsing error in MediaWiki
Action format v2 responses for redirects, and
and issue with content_format not being updated
when the source payload sets a content_model.

Update event schema to 1.2.0 for source/sink
streams to align with the upstream producer
configuration.

Bug: T368746
Change-Id: Id5b82487cdde76b48018d61095d6bc96829e5a10