airflow: define the kerberos keytab secret at all times
Because we don't have an easy way to selectively mount volumes
(as they are defined in `values.yaml`), we define the keytab secret no
matter if kerberos is enabled or not.

This avoids encoding additional complexity in the chart, as all
instances will have kerberos enabled eventually.

Similar to https://gerrit.wikimedia.org/r/c/operations/deployment-charts/+/1087459

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: I65c212d4887f340b5fc8b042e067497e21a6ce3a