changeprop-jobqueue: add latency-sensitive upload jobs rules
This adds a dedicated rule for each of AssembleUploadChunks,
PublishStashedFile, and UploadFromUrl, under the new
latency_sensitive_jobs_config.

To minimize changeprop restarts, the new rules are enabled from the
start. This does open a brief window where the low-traffic and new rules
can execute the same job concurrently (since deduplication is per-rule
in changeprop). In practice, any job that's meant to be idempotent for
the purposes of allowing retries on failure should also be tolerant of
this case, in the absence of a perfect failure detector.

Bug: T379035
Change-Id: I4d6bfe2b8e537aec9e1558c494b1ce966cedb114