changeprop-jobqueue: bump max.poll.interval.ms to 2h
Over the last 24h, running with this set to 1h seems to be highly
effective in reducing timeout-related rebalances: we've only seen one.

Of course, it would be nice to see none at all, so let's try to raise
the poll interval a bit more, starting by doubling to 2h.

Bug: T356241
Change-Id: I77cde4d40ca5f1ceefad8b7b5bd66c3f230efd22