airflow-analytics-test: deploy the scheduler and kerberos components
We're experimenting with migrating the airflow-analytics-test scheduler
to Kubernetes, meaning that after this migration, _all_ components will
be running in containers.

We thus have to enable the kerberos and scheduler components, and we can
remove the network policy allowing the webserver to reach the scheduler
on an-test-client1002.

For now though, we keep using the DB on an-db1001, as this component will
be the last to be migrated.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T380284
Change-Id: I340df8deed626c582c7c1cba7168d14eae84e25a