Enable pod-scoped "external services" network policies for airflow
We sould like to have  Calico Network Policies scoped at the pod level,
to have a finer-grain control over who gets access to what.
This adds a component webserver, scheduler etc. label selector in
the generated Calico Network Policy.
It is currently in progress and shall be tested on the deployed test
environment

Bug: T377926
Change-Id: Id318ab1c6d88fff029b4cb70fe0d5c93c5ab6200