datahub: add datahub production index prefix
We currently share our opensearch cluster between production and staging,
When restoring the opensearch indices for the production,
we noticed that the indices of the staging deployment appeared to have been wiped,
as well as those of the production server.
We feel that this is likely caused by an error in the code that deletes all indices
beginning with the configured prefix. In our case,
the production deployment did not have a prefix, so it deleted them all,
including those with the staging_ prefix.

This patch introduces a production_ prefix for the production deployment indices,
which shall hopefully prevent this from happening again.

Bug: T377814
Change-Id: Ic980d718d1b119bbb8e88b37bd8b463b6659819f