modules: add health checks to the mesh's _tcp_cluster config
We are currently using the _tcp_cluster config to allow Tegola
to proxy to a set of Postgres sql read replicas on the Maps
cluster. As far as I can understand without any health check
config, a target host can go down and keep receiving TCP connection
attempts:

https://www.envoyproxy.io/docs/envoy/v1.23.12/api-v3/config/cluster/v3/cluster.proto

> If no configuration is specified no health checking will be done and all
> cluster members will be considered healthy at all times.

With an optional health check config we should be able
to leverage the default active checking capability of envoy
to remove an endpoint from the LB if not responding to TCP
sockets (for example, hw failure or reboots).

I've set the new health check config to be optional so we can
easily test it on Tegola staging/production, long term we may
want to set a good default.

Bug: T322647
Change-Id: I4931af476ef180202dc3d47e7f5c208bfced5828