modules: remove tpl() usage in base:helper's resourcesDataChecksum
The include() function already evaluates/renders a teplate, and
its result being passed to tpl() raises errors for config-maps
related to ServiceRunner configs, where we have {{ something }}
as well (so include generates the service-runner config correctly,
but then tpl() misinterprets those values like they where part
of the template). The call to tpl() seems unnecessary, so this
change removes it.

Change-Id: If59d74de2db044dd72fd335f519c1faef99b1c52