services: use external_services for maps read replicas
Reduce the IP hardcoded in the Tegola's config leveraging
the external_services config. We could also exclude for safety
the maps masters from the rotation of the read traffic,
Tegola and Kartotherian are meant only to use the read replicas
and preventing any accidental write to the master may
be good long term.

Change-Id: I97253e2ad535a731c5197e657aee291bcd8ffdd6