ceph-csi: remove un-necessary network policies allowing kube api egress
These network policies are both un-necessary and harmful. Un-necessary
because any pod running in the `kube-system` namespace can talk to
anything, and harmful, because by applying a single network policy, the
iptables policy for the associated pod becomes `DROP` and all other
network accesses are denied.

By not specifying any network policy, we then restore access to Ceph and
kube-dns.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T381264
Change-Id: Ic91d39729c8c680f7e04505b8288b658f0d94458