mw-videoscaler: add dummy value for timestamp
We don't require the use of this timestamp for mercurius to roll
versions on new image versions. We use the mediawiki image timestamp
as our identifier, so the timestamp is superfluous for our
purposes. Our use of `now` for the timestamp creates a diff on every
run. This is confusing and messy, for the short-term, use a stub value
and in future we will refactor this behaviour

Bug: T371700
Change-Id: I4067015bf81226e77e874002b8b953e5ea4dca15