mediawiki: add rsyslog container to mercurius
This adds rsyslog to mercurius job's pod template, which is necessary
for logstash-bound events sent to localhost:10514 to be forwarded to
Kafka for ingestion.

While the rsyslog configuration _does_ contain things we don't need in
the mercurius case (e.g., php-fpm log processing), I believe that should
be fine (indeed, this seems to be the case for mwscript already).

To allow an update to the job template in the absence of an image
version change, also bumps the release generation.

Bug: T382517
Change-Id: Iacfaa72a2dc0c74c1e0c833086c8f206e9a499ac