mediawiki: add mercurius release generation token
The mercurius.generation helmfile value, if set, is now appended to the
release ID used in constructing the job name and in both the config and
release control file.

This is useful for deploying helmfile-only changes affecting the job
template (which is immutable) in the absence of an image version change.

Since this increases the (already significant) risk of the truncated job
name colliding between instances for similar-prefix mediawiki jobs, we
now use a *truncated digest* of the release name in the job name.

The *full* release name (with generation suffix) is still used as the
release ID in the config and control file for clarity.

Important points of note:
1. This is not an automatic fix for helmfile-only deployments, as the
   user must still know that the generation value must be set.
2. Due to the change in job name, care must be taken in deploying this
   change: when applied for the first time, generation must be set (as
   it is here), lest we end up with duplicate jobs that consider
   themselves active.

Bug: T382630
Change-Id: I06123307be4b7102465d2d8cd2b235870c108e33