shellbox-video: scale down
The current count of shellbox-video instances was upped to handle a
migration where we overwhelmed shellbox-video a little due to
changeprop being a bit unpredictable in load allocation. Now that
we're on mercurius/mw-videoscaler, we are relatively certain about our
max being (currently) 20 jobs. However, during migrations there is a
chance of us going over this limit so set the limit to 30 for
now. Also add quite a few notes to explain this logic.

Also make mercurius.workers explicit in mw-videoscaler, even if this
is the default used in the mediawiki values file.

Bug: T383317
Change-Id: Ifce7b644491d729df1e040f771ee8563e065935d