airflow: Allow specific task pods to access the kube-api
We are starting to use the KubernetesPodOperator to allow certain tasks
to launch pods with custom images.

In order to do this, we will need to allow these tasks to acccess the
Kubernetes API.

Rather than allowing all task pods to access the Kubernetes API, we will
only apply the related networkpolicy to task-pods that also have a label
of 'operator' set to one of a certain list.

Since we also have the spark-operator active, we wish to include the
Airflow KubernetesSparkOperator.

Bug: T383430
Bug: T318712
Change-Id: I3219bf571f7475e4944fa2a933c2bf2bd392b24b